AlbumArt Service (#42)
* feat : profile 별 파일 분리

* feat : 기존 user 클래스 경로 변경으로 인한 import 경로 수정

* feat(docker) : DB 접속정보 수정

* feat : user 서비스 추가

* feat : User 관련 파일 경로 재설정

* feat : test 폴더 삭제

* feat : user 삭제 soft, hard 로 다시 구현

* feat : user service

* feat(ignore) : 로컬전용 파일 ignore 추가

* feat : main branch 와 경로 일치화

* feat : 내정보 조회 관련 security 접근권한 추가 및 내정보 조회 url 수정

* feat: user 필드 수정

* feat : 임시

* feat : user 내정보 조회 Filter에서 request 에 정보 담는 방식으로 변경

* feat : 어노테이션을 통해 User정보 가져오도록 변경

* config : 서버 실행 에러 발생하는 파일 삭제

* feat : 스웨거 수정

* feat : albumArt Service

* feat : albumArt Service

* feat : albumArt detail, create

* feat : albumart detail

* feat : 임시 accesstoken 발급 로직

* feat : albumart detail v1

* feat : albumArt service

* feat : 내정보조회 Post -> Get 변경

* feat : QClass 삭제

* feat : user service

* feat : albumart service

* feat : albumart filter

* feat : album Art service

* feat : album Art query dsl

* feat : albumart 전체조회 수정

* feat : albumart service

* feat : merge

* feat : albumart create

* feat : albumart contact

* feat : albumart contact

* feat : albumart contact

* feat : albumart type change

* feat : albumart request

* feat : filter pk id로 변경, initData 에서 user 객체 nullpoint 해결

* feat : albumart service

* feat : initData debug

+++ src/main/java/com/study/palette/module/albumArt/service/AlbumArtService.java
@@ -7,13 +7,16 @@ import com.study.palette.module.albumArt.dto.info.AlbumArtCreateRequestDto;
 import com.study.palette.module.albumArt.dto.info.AlbumArtCreateResponseDto;
 import com.study.palette.module.albumArt.dto.info.AlbumArtDetailResponseDto;
 import com.study.palette.module.albumArt.dto.info.AlbumArtResponseDto;
-import com.study.palette.module.albumArt.dto.license.AlbumArtLicenseInfoWithIdDto;
 import com.study.palette.module.albumArt.dto.info.AlbumArtUpdateReqeustDto;
+import com.study.palette.module.albumArt.dto.query.FindAlbumArtQuery;
+import com.study.palette.module.albumArt.entity.AlbumArtContact;
 import com.study.palette.module.albumArt.entity.AlbumArtInfo;
 import com.study.palette.module.albumArt.entity.AlbumArtLicenseInfo;
+import com.study.palette.module.albumArt.entity.AlbumArtRequest;
 import com.study.palette.module.albumArt.exception.AlbumArtErrorCode;
 import com.study.palette.module.albumArt.exception.AlbumArtException;
 import com.study.palette.module.albumArt.repository.AlbumArtRepository;
+import com.study.palette.module.albumArt.repository.AlbumArtRequestRepository;
 import com.study.palette.module.user.entity.User;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -21,7 +24,9 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.time.LocalDate;
 import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
 import java.util.stream.Collectors;
 
@@ -29,25 +34,30 @@ import java.util.stream.Collectors;
 public class AlbumArtService {
 
     private final AlbumArtRepository albumArtRepository;
+    private final AlbumArtRequestRepository albumArtRequestRepository;
     private final ModelMapper modelMapper;
 
     @Autowired
-    public AlbumArtService(AlbumArtRepository albumArtRepository, ModelMapper modelMapper) {
+    public AlbumArtService(AlbumArtRepository albumArtRepository, AlbumArtRequestRepository albumArtRequestRepository, ModelMapper modelMapper) {
         this.albumArtRepository = albumArtRepository;
+        this.albumArtRequestRepository = albumArtRequestRepository;
         this.modelMapper = modelMapper;
     }
 
     /* AlbumArt 필터 포함 조회*/
     @Transactional(readOnly = true)
-    public PaginationDto<AlbumArtResponseDto> getAlbumArts(Pageable pageable) {
+    public PaginationDto<AlbumArtResponseDto> getAlbumArts(FindAlbumArtQuery query, Pageable pageable) {
         Long count = albumArtRepository.count();
 
         if (count == 0) {
             return PaginationDto.of(new PagingDto(pageable, count), List.of());
         }
 
-        List<AlbumArtResponseDto> artists = albumArtRepository.findAll(pageable).stream()
-                .map(AlbumArtResponseDto::new).collect(Collectors.toList());
+        List<AlbumArtResponseDto> artists = albumArtRepository.findAll(query, pageable)
+                .stream()
+                .map(data -> modelMapper.map(data, AlbumArtResponseDto.class))
+                .collect(Collectors.toList());
+
         PaginationDto<AlbumArtResponseDto> row = PaginationDto.of(new PagingDto(pageable, count), artists);
 
         return row;
@@ -58,36 +68,13 @@ public class AlbumArtService {
     public AlbumArtDetailResponseDto getAlbumArtWithDto(String id) {
         AlbumArtInfo albumArtInfo = albumArtRepository.findById(UUID.fromString(id))
                 .orElseThrow(() -> new AlbumArtException(AlbumArtErrorCode.ALBUM_ART_NOT_FOUND));
-
-        List<AlbumArtLicenseInfoWithIdDto> licenses = albumArtInfo.getAlbumArtLicenseInfo().stream()
-                .map(license -> modelMapper.map(license, AlbumArtLicenseInfoWithIdDto.class)).toList();
-
-//        List<AlbumArtFileResponseDto> albumArtFiles = albumArtInfo.getAlbumArtFile().stream()
-//                .map(data -> modelMapper.map(data, AlbumArtFileResponseDto.class)).toList();
-
-        return new AlbumArtDetailResponseDto(
-                albumArtInfo,
-//                albumArtFiles,
-                licenses);
+        return new AlbumArtDetailResponseDto(albumArtInfo);
     }
 
     /* AlbumArt 등록*/
     @Transactional
     public AlbumArtCreateResponseDto createAlbumArt(AlbumArtCreateRequestDto albumArtCreateRequestDto, User user) {
-        AlbumArtInfo albumArtInfo = albumArtCreateRequestDto.toEntity(user);
-
-        //TODO 파일 구현 후 추가
-//        List<AlbumArtFile> files = albumArtCreateRequestDto.getAlbumArtFiles().stream()
-//                .map(file -> AlbumArtFile.from(file, albumArtInfo))
-//                .toList();
-        List<AlbumArtLicenseInfo> licenses = albumArtCreateRequestDto.getAlbumArtLicenseInfo().stream()
-                .map(license -> AlbumArtLicenseInfo.from(license, albumArtInfo))
-                .toList();
-
-        albumArtInfo.setAlbumArtLicenseInfo(licenses);
-//        albumArtInfo.setAlbumArtFile(files);
-
-        return new AlbumArtCreateResponseDto(albumArtRepository.save(albumArtInfo));
+        return new AlbumArtCreateResponseDto(albumArtRepository.save(albumArtCreateRequestDto.toEntity(user)));
     }
 
     /* AlbumArt 수정*/
@@ -105,10 +92,15 @@ public class AlbumArtService {
 
         // update 전 초기화
         albumArtInfo.getAlbumArtLicenseInfo().clear();
+        albumArtInfo.getAlbumArtContact().clear();
 //        albumArtInfo.getAlbumArtFile().clear(); TODO 파일 구현 후 추가
 
         albumArtUpdateReqeustDto.getAlbumArtLicenseInfo().stream().forEach(license -> {
             albumArtInfo.getAlbumArtLicenseInfo().add(AlbumArtLicenseInfo.from(license, albumArtInfo));
+        });
+
+        albumArtUpdateReqeustDto.getAlbumArtContact().stream().forEach(contact -> {
+            albumArtInfo.getAlbumArtContact().add(AlbumArtContact.from(contact, albumArtInfo));
         });
 
         //TODO 파일 구현 후 추가
@@ -119,11 +111,40 @@ public class AlbumArtService {
         albumArtRepository.save(albumArtInfo);
     }
 
-    /* AlbumArt 삭제*/
+    /*
+    AlbumArt 구매의뢰 하루에 한번만 가능
+    */
     @Transactional
-    public void deleteAlbumArt(String id) {
+    public void createAlbumArtRequest(String id, User user) {
         AlbumArtInfo albumArtInfo = albumArtRepository.findById(UUID.fromString(id))
                 .orElseThrow(() -> new AlbumArtException(AlbumArtErrorCode.ALBUM_ART_NOT_FOUND));
+
+        //오늘 이미 구매의뢰를 했는지 체크
+        Optional<AlbumArtRequest> albumArtRequest = albumArtRequestRepository.findByAlbumArtInfoAndUserAndCreatedAt(user, albumArtInfo, LocalDate.now());
+
+        if (albumArtRequest.isPresent()) {
+            throw new AlbumArtException(AlbumArtErrorCode.ALBUM_ART_REQUEST_ALREADY_EXISTS);
+        }
+
+        albumArtRequestRepository.save(AlbumArtRequest.builder()
+                .albumArtInfo(albumArtInfo)
+                .user(user)
+                .createAt(LocalDate.now())
+                .build()
+        );
+    }
+
+    /* AlbumArt 삭제*/
+    @Transactional
+    public void deleteAlbumArt(String id, User user) {
+        AlbumArtInfo albumArtInfo = albumArtRepository.findById(UUID.fromString(id))
+                .orElseThrow(() -> new AlbumArtException(AlbumArtErrorCode.ALBUM_ART_NOT_FOUND));
+
+        //본인이 작성한 글인지 체크
+        if (!albumArtInfo.getUser().getId().equals(user.getId())) {
+            throw new AlbumArtException(AlbumArtErrorCode.ALBUM_ART_NOT_YOURS);
+        }
+
         albumArtRepository.delete(albumArtInfo);
     }
 }
